requirements.txt
python-telegram-bot
ibm-watson
